Dialogs

+++ app/favorites/favorites.component.ts
@@ -6,6 +6,8 @@ import { RadListViewComponent } from 'nativescript-telerik-ui/listview/angular';
 import { ObservableArray } from 'tns-core-modules/data/observable-array';
 import { DrawerPage } from '../shared/drawer/drawer.page';
 import {View} from "tns-core-modules/ui/core/view";
+import {Toasty} from "nativescript-toasty";
+import { confirm } from "ui/dialogs";
 @Component({
     selector: 'app-favorites',
     moduleId: module.id,
@@ -32,9 +34,34 @@ export class FavoritesComponent extends DrawerPage implements OnInit {
     }
 
     deleteFavorite(id: number) {
-        this.favoriteservice.deleteFavorite(id)
-            .subscribe(favorites => this.favorites = new ObservableArray(favorites),
-                errmess => this.errMess = errmess);
+        console.log('delete', id);
+
+        let options = {
+            title: "Confirm Delete",
+            message: 'Do you want to delete Dish '+ id,
+            okButtonText: "Yes",
+            cancelButtonText: "No",
+            neutralButtonText: "Cancel"
+        };
+
+        confirm(options).then((result: boolean) => {
+            if(result) {
+
+                this.favorites = null;
+
+                this.favoriteservice.deleteFavorite(id)
+                    .subscribe(favorites => {
+                            const toast = new Toasty("Deleted Dish "+ id, "short", "bottom");
+                            toast.show();
+                            this.favorites = new ObservableArray(favorites);
+                        },
+                        errmess => this.errMess = errmess);
+            }
+            else {
+                console.log('Delete cancelled');
+            }
+        });
+
     }
 
     public onCellSwiping(args: ListViewEventData) {
